demo_apps/chat-demo/requirements.txt
PyJWT>=2.8.0 